use confuse module top level import internally
