+++ confuse/sources.py
@@ -1,6 +1,6 @@
 from __future__ import division, absolute_import, print_function
 
-from .util import BASESTRING
+import confuse
 
 __all__ = ['ConfigSource']
 
@@ -10,7 +10,7 @@ class ConfigSource(dict):
     """
     def __init__(self, value, filename=None, default=False):
         super(ConfigSource, self).__init__(value)
-        if filename is not None and not isinstance(filename, BASESTRING):
+        if filename is not None and not isinstance(filename, confuse.BASESTRING):
             raise TypeError(u'filename must be a string or None')
         self.filename = filename
         self.default = default
