+++ confuse/templates.py
@@ -4,8 +4,7 @@ import os
 import re
 import sys
 
-from . import exceptions as excs
-from . import util
+import confuse
 
 try:
     import enum
@@ -58,7 +57,7 @@ class Template(object):
         try:
             value, _ = view.first()
             return self.convert(value, view)
-        except excs.NotFoundError:
+        except confuse.NotFoundError:
             pass
 
         # Get default value, or raise if required.
@@ -67,7 +66,7 @@ class Template(object):
     def get_default_value(self, key_name='default'):
         if self.default is REQUIRED:
             # Missing required value. This is an error.
-            raise excs.NotFoundError(u"{} not found".format(key_name))
+            raise confuse.NotFoundError(u"{} not found".format(key_name))
         # Missing value, but not required.
         return self.default
 
@@ -89,7 +88,7 @@ class Template(object):
         mismatch rather than a malformed value. In this case, a more
         specific exception is raised.
         """
-        exc_class = excs.ConfigTypeError if type_error else excs.ConfigValueError
+        exc_class = confuse.ConfigTypeError if type_error else confuse.ConfigValueError
         raise exc_class(
             u'{0}: {1}'.format(view.name, message)
         )
@@ -121,7 +120,7 @@ class Number(Template):
     def convert(self, value, view):
         """Check that the value is an int or a float.
         """
-        if isinstance(value, util.NUMERIC_TYPES):
+        if isinstance(value, confuse.NUMERIC_TYPES):
             return value
         else:
             self.fail(
@@ -207,7 +206,7 @@ class String(Template):
     def convert(self, value, view):
         """Check that the value is a string and matches the pattern.
         """
-        if not isinstance(value, util.BASESTRING):
+        if not isinstance(value, confuse.BASESTRING):
             self.fail(u'must be a string', view, True)
 
         if self.pattern and not self.regex.match(value):
@@ -317,12 +316,12 @@ class OneOf(Template):
                     return view.parent.get(next_template)[view.key]
                 else:
                     return view.get(candidate)
-            except excs.ConfigTemplateError:
+            except confuse.ConfigTemplateError:
                 raise
-            except excs.ConfigError:
+            except confuse.ConfigError:
                 pass
             except ValueError as exc:
-                raise excs.ConfigTemplateError(exc)
+                raise confuse.ConfigTemplateError(exc)
 
         self.fail(
             u'must be one of {0}, not {1}'.format(
@@ -349,7 +348,7 @@ class StrSeq(Template):
         self.split = split
 
     def _convert_value(self, x, view):
-        if isinstance(x, util.STRING):
+        if isinstance(x, confuse.STRING):
             return x
         elif isinstance(x, bytes):
             return x.decode('utf-8', 'ignore')
@@ -360,7 +359,7 @@ class StrSeq(Template):
         if isinstance(value, bytes):
             value = value.decode('utf-8', 'ignore')
 
-        if isinstance(value, util.STRING):
+        if isinstance(value, confuse.STRING):
             if self.split:
                 value = value.split()
             else:
@@ -401,11 +400,11 @@ class Pairs(StrSeq):
         try:
             return (super(Pairs, self)._convert_value(x, view),
                     self.default_value)
-        except excs.ConfigTypeError:
+        except confuse.ConfigTypeError:
             if isinstance(x, abc.Mapping):
                 if len(x) != 1:
                     self.fail(u'must be a single-element mapping', view, True)
-                k, v = util.iter_first(x.items())
+                k, v = confuse.iter_first(x.items())
             elif isinstance(x, abc.Sequence):
                 if len(x) != 2:
                     self.fail(u'must be a two-element list', view, True)
@@ -465,12 +464,12 @@ class Filename(Template):
     def resolve_relative_to(self, view, template):
         if not isinstance(template, (abc.Mapping, MappingTemplate)):
             # disallow config.get(Filename(relative_to='foo'))
-            raise excs.ConfigTemplateError(
+            raise confuse.ConfigTemplateError(
                 u'relative_to may only be used when getting multiple values.'
             )
 
         elif self.relative_to == view.key:
-            raise excs.ConfigTemplateError(
+            raise confuse.ConfigTemplateError(
                 u'{0} is relative to itself'.format(view.name)
             )
 
@@ -499,11 +498,11 @@ class Filename(Template):
             except KeyError:
                 if next_relative in template.subtemplates:
                     # we encountered this config key previously
-                    raise excs.ConfigTemplateError((
+                    raise confuse.ConfigTemplateError((
                         u'{0} and {1} are recursively relative'
                     ).format(view.name, self.relative_to))
                 else:
-                    raise excs.ConfigTemplateError((
+                    raise confuse.ConfigTemplateError((
                         u'missing template for {0}, needed to expand {1}\'s'
                         u'relative path'
                     ).format(self.relative_to, view.name))
@@ -516,16 +515,16 @@ class Filename(Template):
     def value(self, view, template=None):
         try:
             path, source = view.first()
-        except excs.NotFoundError:
+        except confuse.NotFoundError:
             return self.get_default_value(view.name)
 
-        if not isinstance(path, util.BASESTRING):
+        if not isinstance(path, confuse.BASESTRING):
             self.fail(
                 u'must be a filename, not {0}'.format(type(path).__name__),
                 view,
                 True
             )
-        path = os.path.expanduser(util.STRING(path))
+        path = os.path.expanduser(confuse.STRING(path))
 
         if not os.path.isabs(path):
             if self.cwd is not None:
@@ -610,9 +609,9 @@ def as_template(value):
         return Integer()
     elif isinstance(value, int):
         return Integer(value)
-    elif isinstance(value, type) and issubclass(value, util.BASESTRING):
+    elif isinstance(value, type) and issubclass(value, confuse.BASESTRING):
         return String()
-    elif isinstance(value, util.BASESTRING):
+    elif isinstance(value, confuse.BASESTRING):
         return String(value)
     elif isinstance(value, set):
         # convert to list to avoid hash related problems
